% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettycov.R
\name{pretty_cov_plot}
\alias{pretty_cov_plot}
\title{Pretty track plot}
\usage{
pretty_cov_plot(cvg, parts, target, design = NULL, alpha = FALSE, ...)
}
\arguments{
\item{cvg}{Coverage as GRanges or GRangesList}

\item{parts}{Annotation from `superintronic::collect_parts()`}

\item{target}{Region of interest as GRanges or data.frame}

\item{alpha}{(Make low coverage regions transparent?) Default is FALSE}

\item{...}{Other options passed to `patchwork::wrap_plots()`}
}
\description{
Pretty track plot
}
